Continue key colors by creating shared methods
